Reintroduce copy_from
