+++ src/renderer/effect.rs
@@ -5,7 +5,3 @@ pub use fog_effect::*;
 mod fxaa_effect;
 #[doc(inline)]
 pub use fxaa_effect::*;
-
-mod copy_effect;
-#[doc(inline)]
-pub use copy_effect::*;
